+++ source/GreenEnergyHub.Charges/source/GreenEnergyHub.Charges.Infrastructure/Contracts/Internal/ChargeLinksCommandRejected/ChargeLinksCommandRejectedInboundMapper.cs
@@ -16,7 +16,6 @@ using System.Collections.Generic;
 using System.Linq;
 using Energinet.DataHub.Core.Messaging.Protobuf;
 using Energinet.DataHub.Core.Messaging.Transport;
-using Google.Protobuf.Collections;
 using GreenEnergyHub.Charges.Core.DateTime;
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeLinksCommands;
 using GreenEnergyHub.Charges.Domain.Dtos.ChargeLinksRejectionEvents;
@@ -40,13 +39,14 @@ namespace GreenEnergyHub.Charges.Infrastructure.Contracts.Internal.ChargeLinksCo
                     ConvertDocument(chargeLinksCommandRejected.ChargeLinksCommand.Document),
                     chargeLinksCommandRejected.ChargeLinksCommand.ChargeLinks.Select(ConvertChargeLink)
                         .ToList()),
-                ConvertValidationRuleIdentifiers(chargeLinksCommandRejected.FailedValidationRuleIdentifiers));
+                ConvertValidationError(chargeLinksCommandRejected.ValidationErrors));
         }
 
-        private IEnumerable<ValidationRuleIdentifier> ConvertValidationRuleIdentifiers(
-            RepeatedField<ValidationRuleIdentifierContract> failedValidationRuleIdentifierContracts)
+        private IEnumerable<ValidationError> ConvertValidationError(
+            IEnumerable<ValidationErrorContract> validationErrorContracts)
         {
-            return failedValidationRuleIdentifierContracts.Select(x => (ValidationRuleIdentifier)x);
+            return validationErrorContracts.Select(x =>
+                new ValidationError((ValidationRuleIdentifier)x.ValidationRuleIdentifier, x.TriggeredBy));
         }
 
         private static DocumentDto ConvertDocument(Document document)