+++ source/Energinet.Charges.Libraries/source/Energinet.DataHub.Charges.Clients/documents/release-notes/release-notes.md
@@ -1,5 +1,13 @@
 # Energinet.DataHub.Charges.Clients Release notes
 
+## Version 1.0.24
+
+Removed local settings to the unused resources earlier used to request the creation of messages
+
+## Version 1.0.23
+
+Updated Nuget packages
+
 ## Version 1.0.22
 
 `ChargeLinkClient` uses `JsonStringEnumConverter` to handle `enum` passed as string